Merge two images together and unmerge them
Not the prettiest code I'll admit but it was pretty fun to write.

The dimensions for the hidden image are hidden in the first pixels
of the merged image. The first byte holds in its last 4 bits (nibble)
the bytes each dimension takes up (0-15). If it has a value of 3
for example, then you will find the width in the next 3 bytes
(in their lower nibbles) and height in the 3 after those.